+++ tidy3d/components/source.py
@@ -35,6 +35,8 @@ DATA_SPAN_TOL = 1e-8
 CHEB_GRID_WIDTH = 1.5
 # Number of frequencies in a broadband source above which to issue a warning
 WARN_NUM_FREQS = 20
+# how many units of ``twidth`` from the ``offset`` until a gaussian pulse is considered "off"
+END_TIME_FACTOR_GAUSSIAN = 10
 
 
 class SourceTime(AbstractTimeDependence):
@@ -78,6 +80,10 @@ class SourceTime(AbstractTimeDependence):
     def frequency_range(self, num_fwidth: float = 4.0) -> FreqBound:
         """Frequency range within plus/minus ``num_fwidth * fwidth`` of the central frequency."""
 
+    @abstractmethod
+    def end_time(self) -> float | None:
+        """Time after which the source is effectively turned off / close to zero amplitude."""
+
 
 class Pulse(SourceTime, ABC):
     """A source time that ramps up with some ``fwidth`` and oscillates at ``freq0``."""
@@ -99,6 +105,11 @@ class Pulse(SourceTime, ABC):
         "pulse in units of 1 / (``2pi * fwidth``).",
         ge=2.5,
     )
+
+    @property
+    def twidth(self) -> float:
+        """Width of pulse in seconds."""
+        return 1.0 / (2 * np.pi * self.fwidth)
 
     def frequency_range(self, num_fwidth: float = 4.0) -> FreqBound:
         """Frequency range within 5 standard deviations of the central frequency.
@@ -143,24 +154,32 @@ class GaussianPulse(Pulse):
     def amp_time(self, time: float) -> complex:
         """Complex-valued source amplitude as a function of time."""
 
-        twidth = 1.0 / (2 * np.pi * self.fwidth)
         omega0 = 2 * np.pi * self.freq0
-        time_shifted = time - self.offset * twidth
+        time_shifted = time - self.offset * self.twidth
 
         offset = np.exp(1j * self.phase)
         oscillation = np.exp(-1j * omega0 * time)
-        amp = np.exp(-(time_shifted**2) / 2 / twidth**2) * self.amplitude
+        amp = np.exp(-(time_shifted**2) / 2 / self.twidth**2) * self.amplitude
 
         pulse_amp = offset * oscillation * amp
 
         # subtract out DC component
         if self.remove_dc_component:
-            pulse_amp = pulse_amp * (1j + time_shifted / twidth**2 / omega0)
+            pulse_amp = pulse_amp * (1j + time_shifted / self.twidth**2 / omega0)
         else:
             # 1j to make it agree in large omega0 limit
             pulse_amp = pulse_amp * 1j
 
         return pulse_amp
+
+    def end_time(self) -> float | None:
+        """Time after which the source is effectively turned off / close to zero amplitude."""
+
+        # TODO: decide if we should continue to return an end_time if the DC component remains
+        # if not self.remove_dc_component:
+        #     return None
+
+        return self.offset * self.twidth + END_TIME_FACTOR_GAUSSIAN * self.twidth
 
 
 class ContinuousWave(Pulse):
@@ -190,6 +209,10 @@ class ContinuousWave(Pulse):
         amp = 1 / (1 + np.exp(-time_shifted / twidth)) * self.amplitude
 
         return const * offset * oscillation * amp
+
+    def end_time(self) -> float | None:
+        """Time after which the source is effectively turned off / close to zero amplitude."""
+        return None
 
 
 class CustomSourceTime(Pulse):
@@ -329,6 +352,20 @@ class CustomSourceTime(Pulse):
         amp = self.amplitude
 
         return offset * oscillation * amp * envelope
+
+    def end_time(self) -> float | None:
+        """Time after which the source is effectively turned off / close to zero amplitude."""
+
+        if self.source_time_dataset is None:
+            return None
+
+        data_array = self.source_time_dataset.values
+
+        t_coords = data_array.coords["t"]
+        source_is_non_zero = ~np.isclose(abs(data_array), 0)
+        t_non_zero = t_coords[source_is_non_zero]
+
+        return np.max(t_non_zero)
 
 
 SourceTimeType = Union[GaussianPulse, ContinuousWave, CustomSourceTime]